Do not print to stdout, instead use stderr.
This fixes error handling in the application. The application no
longer prints to stdout. Everything error and usage related is printed
to stderr.
Furthermore, the application returns EXIT_SUCCESS only if everything
went fine. Even if the target is "all" and only a single device cannot
be set, the application will still continue the operation on all
remaining devices but will return EXIT_FAILURE to signal, that the
operation did not perform completely.
EXIT_SUCCESS signals that *everything* went fine.